fix: change tron opreation trait get_value return type

+++ wallet-chain-interact/src/tron/operations/mod.rs
@@ -20,7 +20,8 @@ pub trait TronTxOperation<T> {
 
     fn get_to(&self) -> String;
 
-    fn get_value(&self) -> i64;
+    // unit is trx
+    fn get_value(&self) -> f64;
 }
 
 pub trait TronSimulateOperation {
